--- src/main/java/com/giftforyoube/global/config/Config.java
@@ -1,4 +1,0 @@
-package com.giftforyoube.global.config;
-
-public class Config {
-}
