+++ src/main/java/com/giftforyoube/global/config/RestTemplateConfig.java
@@ -1,4 +1,4 @@
-package com.sparta.giftforyou.global.config;
+package com.giftforyoube.global.config;
 
 import org.springframework.boot.web.client.RestTemplateBuilder;
 import org.springframework.context.annotation.Bean;
