+++ src/main/java/com/giftforyoube/global/jwt/JwtAuthenticationFilter.java
@@ -1,10 +1,10 @@
-package com.sparta.giftforyou.global.jwt;
+package com.giftforyoube.global.jwt;
 
 import com.fasterxml.jackson.databind.ObjectMapper;
-import com.sparta.giftforyou.domain.user.dto.LoginRequestDto;
-import com.sparta.giftforyou.domain.user.dto.MsgResponseDto;
-import com.sparta.giftforyou.global.security.UserDetailsImpl;
-import com.sparta.giftforyou.global.security.UserDetailsServiceImpl;
+import com.giftforyoube.user.dto.LoginRequestDto;
+import com.giftforyoube.user.dto.MsgResponseDto;
+import com.giftforyoube.global.security.UserDetailsImpl;
+import com.giftforyoube.global.security.UserDetailsServiceImpl;
 import jakarta.servlet.FilterChain;
 import jakarta.servlet.ServletException;
 import jakarta.servlet.http.HttpServletRequest;
