--- src/main/java/com/giftforyoube/global/security/Security.java
@@ -1,4 +1,0 @@
-package com.giftforyoube.global.security;
-
-public class Security {
-}
