+++ src/main/java/com/giftforyoube/user/dto/KakaoUserInfoDto.java
@@ -1,4 +1,4 @@
-package com.sparta.giftforyou.domain.user.dto;
+package com.giftforyoube.user.dto;
 
 import com.fasterxml.jackson.annotation.JsonIgnoreProperties;
 import lombok.Getter;
