--- src/main/java/com/giftforyoube/user/dto/UserRequestDto.java
@@ -1,7 +1,0 @@
-package com.giftforyoube.user.dto;
-
-import lombok.Getter;
-
-@Getter
-public class UserRequestDto {
-}
