+++ src/main/java/com/giftforyoube/user/entity/User.java
@@ -1,9 +1,6 @@
 package com.giftforyoube.user.entity;
 
-import jakarta.persistence.Entity;
-import jakarta.persistence.GeneratedValue;
-import jakarta.persistence.GenerationType;
-import jakarta.persistence.Id;
+import jakarta.persistence.*;
 import lombok.Getter;
 import lombok.NoArgsConstructor;
 
@@ -11,8 +8,41 @@ import lombok.NoArgsConstructor;
 @Getter
 @NoArgsConstructor
 public class User {
-
     @Id
     @GeneratedValue(strategy = GenerationType.IDENTITY)
     private Long id;
+
+    @Column(unique = true)
+    private String email;
+
+    @Column
+    private String password;
+
+    @Column
+    private String nickname;
+
+    @Column(unique = true)
+    private String phoneNumber = "";
+
+    @Column
+    private Long kakaoId = 0L;
+
+    public User(String email, String password, String nickname, String phoneNumber) {
+        this.email = email;
+        this.password = password;
+        this.nickname = nickname;
+        this.phoneNumber = phoneNumber;
+    }
+
+    public User(String email, String password, String nickname, Long kakaoId) {
+        this.email = email;
+        this.password = password;
+        this.nickname = nickname;
+        this.kakaoId = kakaoId;
+    }
+
+    public User kakaoIdUpdate(Long kakaoId) {
+        this.kakaoId = kakaoId;
+        return this;
+    }
 }
