+++ src/main/java/com/giftforyoube/user/repository/UserRepository.java
@@ -1,8 +1,12 @@
 package com.giftforyoube.user.repository;
 
-import com.giftforyoube.donation.entity.Donation;
 import com.giftforyoube.user.entity.User;
 import org.springframework.data.jpa.repository.JpaRepository;
 
+import java.util.Optional;
+
 public interface UserRepository extends JpaRepository<User, Long> {
+    Optional<User> findByEmail(String email);
+    Optional<User> findByPhoneNumber(String phoneNumber);
+    Optional<User> findByKakaoId(Long kakaoId);
 }
