--- src/main/java/com/sparta/giftforyou/GiftForYouApplication.java
@@ -1,13 +1,0 @@
-package com.sparta.giftforyou;
-
-import org.springframework.boot.SpringApplication;
-import org.springframework.boot.autoconfigure.SpringBootApplication;
-
-@SpringBootApplication
-public class GiftForYouApplication {
-
-	public static void main(String[] args) {
-		SpringApplication.run(GiftForYouApplication.class, args);
-	}
-
-}
