--- src/main/java/com/sparta/giftforyou/domain/user/repository/UserRepository.java
@@ -1,12 +1,0 @@
-package com.sparta.giftforyou.domain.user.repository;
-
-import com.sparta.giftforyou.domain.user.entity.User;
-import org.springframework.data.jpa.repository.JpaRepository;
-
-import java.util.Optional;
-
-public interface UserRepository extends JpaRepository<User, Long> {
-    Optional<User> findByEmail(String email);
-    Optional<User> findByPhoneNumber(String phoneNumber);
-    Optional<User> findByKakaoId(Long kakaoId);
-}
