--- src/test/java/com/sparta/giftforyou/GiftForYouApplicationTests.java
@@ -1,13 +1,0 @@
-package com.sparta.giftforyou;
-
-import org.junit.jupiter.api.Test;
-import org.springframework.boot.test.context.SpringBootTest;
-
-@SpringBootTest
-class GiftForYouApplicationTests {
-
-	@Test
-	void contextLoads() {
-	}
-
-}
